refactor(eclipse): extract interface InlineCompletionItemRenderer.
