+++ clients/eclipse/plugin/src/com/tabbyml/tabby4eclipse/inlineCompletion/renderer/InlineCompletionItemTextPainter.java
@@ -1,4 +1,4 @@
-package com.tabbyml.tabby4eclipse.inlineCompletion;
+package com.tabbyml.tabby4eclipse.inlineCompletion.renderer;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -27,71 +27,27 @@ import org.eclipse.swt.graphics.Point;
 import org.eclipse.swt.widgets.Display;
 
 import com.tabbyml.tabby4eclipse.Logger;
+import com.tabbyml.tabby4eclipse.inlineCompletion.InlineCompletionItem;
 
-public class InlineCompletionRenderer {
-	private Logger logger = new Logger("InlineCompletionRenderer");
-	private Map<ITextViewer, InlineCompletionItemPainter> painters = new HashMap<>();
-	private ITextViewer currentTextViewer = null;
-	private InlineCompletionItem currentCompletionItem = null;
-	private String currentViewId = null;
-	private Long currentDisplayAt = null;
+public class InlineCompletionItemTextPainter implements IInlineCompletionItemRenderer {
+	private Logger logger = new Logger("InlineCompletionRenderer.InlineCompletionItemTextPainter");
+	private Map<ITextViewer, GhostTextPainter> painters = new HashMap<>();
 
-	public void show(ITextViewer viewer, InlineCompletionItem item) {
-		if (currentTextViewer != null) {
-			getPainter(currentTextViewer).update(null);
-		}
-		currentTextViewer = viewer;
-		currentCompletionItem = item;
-		getPainter(viewer).update(item);
-		
-		currentDisplayAt = System.currentTimeMillis();
-		
-		String completionId = "no-cmpl-id";
-		if (item.getEventId() != null && item.getEventId().getCompletionId()!= null) {
-			completionId = item.getEventId().getCompletionId().replace("cmpl-", "");
-		}
-		currentViewId = String.format("view-%s-at-%d", completionId, currentDisplayAt);
+	@Override
+	public void updateInlineCompletionItem(ITextViewer textViewer, InlineCompletionItem item) {
+		getPainter(textViewer).update(item);
 	}
 
-	public void hide() {
-		if (currentTextViewer != null) {
-			getPainter(currentTextViewer).update(null);
-			currentTextViewer = null;
-		}
-		currentCompletionItem = null;
-		currentViewId = null;
-		currentDisplayAt = null;
-	}
-
-	public ITextViewer getCurrentTextViewer() {
-		return currentTextViewer;
-	}
-
-	public InlineCompletionItem getCurrentCompletionItem() {
-		return currentCompletionItem;
-	}
-	
-	public String getCurrentViewId() {
-		return currentViewId;
-	}
-	
-	public Long getCurrentDisplayedTime() {
-		if (currentDisplayAt != null) {
-			return System.currentTimeMillis() - currentDisplayAt;
-		}
-		return null;
-	}
-
-	private InlineCompletionItemPainter getPainter(ITextViewer viewer) {
-		InlineCompletionItemPainter painter = painters.get(viewer);
+	private GhostTextPainter getPainter(ITextViewer viewer) {
+		GhostTextPainter painter = painters.get(viewer);
 		if (painter == null) {
-			painter = new InlineCompletionItemPainter(viewer);
+			painter = new GhostTextPainter(viewer);
 			painters.put(viewer, painter);
 		}
 		return painter;
 	}
 
-	private class InlineCompletionItemPainter implements IPainter, PaintListener {
+	private class GhostTextPainter implements IPainter, PaintListener {
 		private ITextViewer viewer;
 
 		private InlineCompletionItem item;
@@ -103,7 +59,7 @@ public class InlineCompletionRenderer {
 		private List<GlyphMetrics> modifiedGlyphMetrics = new ArrayList<>();
 		private List<Consumer<GC>> paintFunctions = new ArrayList<>();
 
-		public InlineCompletionItemPainter(ITextViewer viewer) {
+		public GhostTextPainter(ITextViewer viewer) {
 			this.viewer = viewer;
 			getDisplay().syncExec(() -> {
 				((ITextViewerExtension2) this.viewer).addPainter(this);
@@ -331,14 +287,15 @@ public class InlineCompletionRenderer {
 					Point location = widget.getLocationAtOffset(offset);
 					gc.drawString(textWithTabs.text, location.x + tabWidth, location.y);
 				});
-				
+
 			} else {
 				// otherwise, draw the ghost text, and move target char after the ghost text
 				String targetChar = widget.getText(targetOffset, targetOffset);
 				StyleRange originStyleRange;
 				if (widget.getStyleRangeAtOffset(targetOffset) != null) {
 					originStyleRange = widget.getStyleRangeAtOffset(targetOffset);
-					logger.debug("Find origin StyleRange:" + originStyleRange.start + " -> " + originStyleRange.metrics);
+					logger.debug(
+							"Find origin StyleRange:" + originStyleRange.start + " -> " + originStyleRange.metrics);
 				} else {
 					originStyleRange = new StyleRange();
 					originStyleRange.start = targetOffset;
