+++ clients/eclipse/plugin/src/com/tabbyml/tabby4eclipse/lsp/protocol/ClientInfo.java
@@ -7,7 +7,7 @@ public class ClientInfo {
 
 	public ClientInfo() {
 	}
-	
+
 	public String getName() {
 		return name;
 	}
