+++ clients/eclipse/plugin/src/com/tabbyml/tabby4eclipse/lsp/protocol/ClientProvidedConfig.java
@@ -8,7 +8,7 @@ public class ClientProvidedConfig {
 
 	public ClientProvidedConfig() {
 	}
-	
+
 	public ServerConfig getServer() {
 		return server;
 	}
